If someone PMs the bot, they likely want a reply.
"Directed" appears to be an umbrella encompassing all messages which a user
sends with the explicit intent of getting the bot to reply.

The other way to fix the "questions in PM without the bot's nick don't get
replies" bug would be to check for PM in every plugin alongside checking
directed, but that seems silly.